Moved LocationInformationInterface to Data namespace
