+++ app/code/Magento/CatalogInventory/Api/InventoryCriteriaInterface.php
@@ -22,8 +22,8 @@ interface InventoryCriteriaInterface extends StockCriteriaInterface
     /**
      * Sets location filter, to retrieve inventory records based on location
      *
-     * @param LocationInformationInterface $location
+     * @param Data\LocationInformationInterface $location
      * @return $this
      */
-    public function setLocationFilter(LocationInformationInterface $location);
+    public function setLocationFilter(Data\LocationInformationInterface $location);
 }
