+++ app/code/Magento/CatalogInventory/Api/InventoryIndexRecordCriteriaInterface.php
@@ -28,10 +28,10 @@ interface InventoryIndexRecordCriteriaInterface extends \Magento\Framework\Api\C
     /**
      * Sets location filter, to retrieve inventory records based on location
      *
-     * @param LocationInformationInterface $location
+     * @param Data\LocationInformationInterface $location
      * @return $this
      */
-    public function setLocationFilter(LocationInformationInterface $location);
+    public function setLocationFilter(Data\LocationInformationInterface $location);
 
     /**
      * Add scope filter to collection
